ecosplit/backend/requirements.txt
Flask
flask-cors
python-dotenv
pandas
scikit-learn
joblib
openai
vertexai
google-cloud-aiplatform
google-cloud-storage